testing b and b for knapsack; works fine, solve time is pb dependant, 15s for 2.5M nodes out of 1B visited (0.002%)

+++ knapsack_hamiltonian_relaxation.py
@@ -1,7 +1,7 @@
 import random
 class Knapsack:
     
-    def __init__(self, weights: list[int], values: list[int], max_weight: int):
+    def __init__(self, weights: list[int], values: list[float], max_weight: int):
         self.weights = weights
         self.values = values
 
@@ -23,7 +23,7 @@ class Knapsack:
         return Knapsack(weights=weights, values=values, max_weight=max_weight)
 
     
-    def solve_dynamic_programming(self, full_result: bool = False) -> int:
+    def solve_dynamic_programming(self, full_result: bool = False) -> float:
 
         vs = self.values
         ws = self.weights
